Changed the type of traffic flows from int to double.

+++ Algorithms/TrafficAssignment/ObjectiveFunctions/UserEquilibrium.h
@@ -13,7 +13,7 @@ class UserEquilibrium {
   UserEquilibrium(TravelCostFunctionT function) : travelCostFunction(function) {}
 
   // Returns the weight of edge e, given the traffic flow x on e.
-  int getEdgeWeight(const int e, const int x) const {
+  int getEdgeWeight(const int e, const double x) const {
     return std::round(travelCostFunction(e, x));
   }
